analytics/requirements.txt
supabase==1.0.3
folium==0.14.0
python-dotenv==1.0.0 